requirements.txt
python-telegram-bot==22.0
yt-dlp==2025.4.30
requests==2.31.0
certifi==2024.2.2
python-dotenv==1.0.1 